remove unnecessary unit tests
